+++ gribpcraster/application/manipulation/__init__.py
@@ -1,2 +1,0 @@
-__author__="nappodo"
-__date__ ="$Aug 4, 2011 01:46:32 AM$"