+++ tests/test_api.py
@@ -1,10 +1,10 @@
 import json
 import os
 import unittest
-import gribpcraster.api
+from main import api
 import util.files
-from gribpcraster.config import Configuration
-from gribpcraster.exc.ApplicationException import ApplicationException
+from main.config import Configuration
+from main.exceptions import ApplicationException
 
 
 class TestAPI(unittest.TestCase):
@@ -18,15 +18,15 @@ class TestAPI(unittest.TestCase):
 
     def setUp(self):
 
-        self._command = gribpcraster.api.command()
+        self._command = api.command()
         args_string = '-l ERROR -c /pyg2p_git/execution_templates_devel/eue_t24.json -i /dataset/test_2013330702/EpsN320-2013063000.grb -o /dataset/testdiffmaps/eueT24 -m 10'
-        self._command2 = gribpcraster.api.command(args_string)
+        self._command2 = api.command(args_string)
         self.conf.remove_geopotential('T3999.gph.grb')
         util.files.delete_file(os.path.join(self.conf.geopotentials.data_path, 'T3999.gph.grb'))
 
     def test_API_init(self):
 
-        self.assertIsInstance(self._command, gribpcraster.api.Command)
+        self.assertIsInstance(self._command, api.Command)
         command = self._command.with_cmdpath('a.json')
         self.assertDictEqual(command._d, {'-c': 'a.json', '-l': 'ERROR'})
         command.with_inputfile('0.grb')
@@ -37,7 +37,7 @@ class TestAPI(unittest.TestCase):
                              {'-c': 'a.json', '-i': '0.grb', '-o': '/dataout/test',
                               '-s': '6', '-e': '240', '-m': '10', '-f': '1', '-x': '4', '-l': 'ERROR'})
         self.assertEqual(str(command), 'pyg2p.py -c a.json -e 240 -f 1 -i 0.grb -l ERROR -m 10 -o /dataout/test -s 6 -x 4')
-        self.assertRaises(ApplicationException, gribpcraster.api.run_command, *(command,))
+        self.assertRaises(ApplicationException, api.run_command, *(command,))
 
     def test_API_run(self):
         self.assertDictEqual(self._command2._d,
@@ -46,9 +46,9 @@ class TestAPI(unittest.TestCase):
                               '-o': self.out_1,
                               '-m': '10', '-l': 'ERROR'})
 
-        ret = gribpcraster.api.run_command(self._command2)
+        ret = api.run_command(self._command2)
         self.assertEqual(ret, 0)
-        # TODO test other cases for a coverage of 80% of gribpcraster package
+        # TODO test other cases for a coverage of 80% of main package
         # scipy interpol, grib interpol, construction of intertables for both,
         # two spatial resolutions, two time resolutions...